+++ addons/main/config.cpp
@@ -8,9 +8,9 @@ class CfgPatches
         requiredAddons[] = {};
         units[] = {};
         weapons[] = {};
-        version = "1.0.0";
-        versionStr = "1.0.0";
-        versionAr[] = { 1, 0, 0 };
+        version = "1.0.1";
+        versionStr = "1.0.1";
+        versionAr[] = { 1, 0, 1 };
     };
 };
 
@@ -33,6 +33,8 @@ class CfgFunctions
             };
 
             class publish {};
+            class openBrowser {};
+            class addPauseMenuButton {};
             class collectMarker {};
             class entityId {};
             class colorToHex {};