+++ tools/Build-TestMod.ps1
@@ -2,9 +2,9 @@
 param(
     [switch]$SkipValidation,
     [ValidatePattern('^@cTab Web Companion Test(?: v[0-9]+)?$')]
-    [string]$StageFolderName = "@cTab Web Companion Test v35",
+    [string]$StageFolderName = "@cTab Web Companion Test v36",
     [ValidatePattern('^cTab-Web-Companion-Test(?:-v[0-9]+)?\.zip$')]
-    [string]$ArchiveFileName = "cTab-Web-Companion-Test-v35.zip"
+    [string]$ArchiveFileName = "cTab-Web-Companion-Test-v36.zip"
 )
 
 Set-StrictMode -Version Latest
@@ -53,7 +53,6 @@ try {
 
     $sources = [ordered]@{
         "addons\ctab_web_main.pbo" = ".hemttout\build\addons\ctab_web_main.pbo"
-        "ctab_web_bridge.dll" = "target\i686-pc-windows-msvc\release\ctab_web_bridge.dll"
         "ctab_web_bridge_x64.dll" = "target\x86_64-pc-windows-msvc\release\ctab_web_bridge.dll"
         "ctab-web-companion.exe" = "target\x86_64-pc-windows-msvc\release\ctab-web-companion.exe"
         "mod.cpp" = "distribution\mod.cpp"
@@ -96,7 +95,7 @@ try {
     $manifest = [ordered]@{
         Product = "cTab Web Companion"
         Author = "[GRP9] NiklasKy"
-        Version = "1.0.0-test-v35"
+        Version = "1.0.1-test-v36"
         GeneratedAtUtc = [DateTime]::UtcNow.ToString("o")
         Signed = $false
         BattlEyeApproved = $false